refactor: Parllax 컴포넌트를 styled-components 형태로 리팩토링
