--- FrontEnd/components/Parallax/Parallax.js
@@ -1,70 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
-
-// core components
-import styles from 'public/jss/material-kit-react/components/parallaxStyle.js';
-
-const useStyles = makeStyles(styles);
-
-export default function Parallax(props) {
-  const [transform, setTransform] = useState('translate3d(0,' + 0 + 'px,0)');
-
-  useEffect(() => {
-    let windowScrollTop;
-    if (window.innerWidth >= 768) {
-      windowScrollTop = window.pageYOffset / 3;
-    } else {
-      windowScrollTop = 0;
-    }
-
-    if (window.innerWidth >= 768) {
-      window.addEventListener('scroll', resetTransform);
-    }
-    return function cleanup() {
-      if (window.innerWidth >= 768) {
-        window.removeEventListener('scroll', resetTransform);
-      }
-    };
-  }, []);
-
-  const resetTransform = () => {
-    var windowScrollTop = window.pageYOffset / 3;
-    setTransform('translate3d(0,' + windowScrollTop + 'px,0)');
-  };
-  const { filter, className, children, style, image, small } = props;
-  const classes = useStyles();
-  const parallaxClasses = classNames({
-    [classes.parallax]: true,
-    [classes.filter]: filter,
-    [classes.small]: small,
-    [className]: className !== undefined,
-  });
-  return (
-    <div
-      className={parallaxClasses}
-      style={{
-        ...style,
-        backgroundImage: `linear-gradient(rgba(0,0,0,0.45), rgba(0,0,0,0.45)), url(${image})`,
-        // backgroundSize: 'cover',
-        backgroundSize: '100%',
-        backgroundColor: '#000000',
-        backgroundPosition: 'center',
-        backgroundRepeat: 'no-repeat',
-        transform: transform,
-      }}
-    >
-      {children}
-    </div>
-  );
-}
-
-Parallax.propTypes = {
-  className: PropTypes.string,
-  filter: PropTypes.bool,
-  children: PropTypes.node,
-  style: PropTypes.string,
-  image: PropTypes.string,
-  small: PropTypes.bool,
-};
