+++ FrontEnd/components/Parallax/ParallaxContent.tsx
@@ -1,0 +1,35 @@
+import styled from 'styled-components';
+
+interface ParallaxContentProps {
+  children: React.ReactNode;
+}
+
+const ParallaxContent = ({ children }: ParallaxContentProps) => {
+  return <Container>{children}</Container>;
+};
+
+const Container = styled.div`
+  padding-right: 15px;
+  padding-left: 15px;
+  margin-right: auto;
+  margin-left: auto;
+  width: 100%;
+
+  @media (min-width: 576px) {
+    max-width: 540px;
+  }
+
+  @media (min-width: 768px) {
+    max-width: 720px;
+  }
+
+  @media (min-width: 992px) {
+    max-width: 960px;
+  }
+
+  @media (min-width: 1200px) {
+    max-width: 1140px;
+  }
+`;
+
+export default ParallaxContent;
