+++ FrontEnd/pages/index.tsx
@@ -1,23 +1,117 @@
 import type { NextPage } from 'next';
 import { NextSeo } from 'next-seo';
+import { useEffect, useState } from 'react';
 import styled from 'styled-components';
+import InfiniteScroll from 'react-infinite-scroll-component';
+
+// mui
+import { makeStyles } from '@material-ui/core/styles';
+import { Typography } from '@material-ui/core';
 
 // core components
 import Header from 'components/Header/Header';
+import HeaderLinks from 'components/Header/HeaderLinks';
 import Footer from 'components/Footer/Footer.js';
 import GridContainer from 'components/Grid/GridContainer';
 import GridItem from 'components/Grid/GridItem';
-import Parallax from 'components/Parallax/Parallax.js';
+import Parallax from 'components/Parallax/Parallax';
+import ParallaxContent from 'components/Parallax/ParallaxContent';
+import PhotoFeed from 'components/Feed/PhotoFeed';
+import CirCularLoader from 'components/Loader/CirCularLoader';
 
-// sections for this page
-import HeaderLinks from 'components/Header/HeaderLinks';
-import SectionInfiniteList from 'components/Sections/SectionInfiniteList';
-
-// Image
+// Icons
 import ScrollGuide from 'components/ScrollGuide/ScrollGuide';
+
+// styles
+import styles from 'public/jss/material-kit-react/views/componentsSections/basicsStyle.js';
+
+// Hooks
+import useBasedInfo from 'Hooks/api/useBasedInfo';
+import useTotalBasedCnt from 'Hooks/api/useTotalBasedCnt';
+
+// util
+import { generateShuffledArr } from 'util/arrUtil';
+
+// models
+import { _iBasedInfoReqParams } from 'models/api/goCamping/basedInfo';
+
+const useStyles = makeStyles(styles);
+
+const EndMessageComponent = styled(Typography)`
+  font-weight: 600;
+  margin-top: 36px;
+  text-align: center;
+`;
+
+const NUM_OF_ROWS = 10;
 
 const Home: NextPage = (props) => {
   const { ...rest } = props;
+  const classes = useStyles();
+
+  // local state
+  const [basedInfoReqParams, setBasedInfoReqParams] = useState<_iBasedInfoReqParams>({
+    pageNo: 0,
+    numOfRows: 0,
+  });
+  const [shuffledPageIdxArr, setShuffledPageIdxArr] = useState([]);
+  const [infBasedList, setInfBasedList] = useState([]);
+
+  const {
+    data: basedInfo,
+    error: basedInfoError,
+    isFetched: basedInfoIsFetched,
+    refetch: getBasedInfo,
+  } = useBasedInfo(basedInfoReqParams);
+
+  const {
+    data: totalBasedCnt,
+    error: totalBasedCntError,
+    isFetching: totalBasedCntIsFetching,
+    isFetched: totalBasedCntIsFetched,
+    refetch: getTotalBasedCnt,
+  } = useTotalBasedCnt();
+
+  // Infinite Scroll의 next 요청
+  const fetchNextData = () => {
+    setBasedInfoReqParams((prev) => ({
+      pageNo: shuffledPageIdxArr[prev.pageNo + 1],
+      numOfRows: NUM_OF_ROWS,
+    }));
+  };
+
+  // Init 하면 basedInfo의 total 갯수를 가져온다.
+  useEffect(() => {
+    getTotalBasedCnt();
+  }, []);
+
+  // 랜덤 페이지 요청을 보내기 위한 셔플 배열 생성
+  useEffect(() => {
+    if (totalBasedCntIsFetched) {
+      const totalPage = Math.ceil(totalBasedCnt / NUM_OF_ROWS);
+      const shuffledPageIdxArr = generateShuffledArr(totalPage).filter((idx) => idx !== totalPage);
+
+      setShuffledPageIdxArr(shuffledPageIdxArr);
+    }
+  }, [totalBasedCntIsFetched]);
+
+  // basedInfo 요청 (parameter가 변경될 때 요청을 보낸다)
+  useEffect(() => {
+    if (basedInfoReqParams.pageNo !== 0) {
+      getBasedInfo();
+    }
+  }, [basedInfoReqParams]);
+
+  // Infinite Based List를 만든다.
+  useEffect(() => {
+    if (basedInfoIsFetched) {
+      if (basedInfo.itemList.length > 0) {
+        const basedItem = basedInfo.itemList;
+
+        setInfBasedList((prev) => prev.concat(basedItem));
+      }
+    }
+  }, [basedInfoIsFetched]);
 
   return (
     <div>
@@ -59,36 +153,35 @@ const Home: NextPage = (props) => {
       </Parallax>
 
       <BodyContainer>
-        <SectionInfiniteList />
+        <div className={classes.sections}>
+          <div className={classes.container}>
+            <InfiniteScroll
+              style={{ overflow: 'hidden' }}
+              dataLength={infBasedList.length}
+              next={fetchNextData}
+              hasMore
+              loader={<CirCularLoader />}
+              endMessage={
+                <EndMessageComponent variant="h4">
+                  더 이상 피드가 없습니다. &#128517;
+                </EndMessageComponent>
+              }
+            >
+              <GridContainer>
+                {infBasedList.map((infBasedItem, idx) => (
+                  <GridItem key={String(idx)} style={{ paddingTop: 16, paddingBottom: 16 }}>
+                    <PhotoFeed basedItem={infBasedItem} />
+                  </GridItem>
+                ))}
+              </GridContainer>
+            </InfiniteScroll>
+          </div>
+        </div>
       </BodyContainer>
       {/* <Footer /> */}
     </div>
   );
 };
-
-const ParallaxContent = styled.div`
-  padding-right: 15px;
-  padding-left: 15px;
-  margin-right: auto;
-  margin-left: auto;
-  width: 100%;
-
-  @media (min-width: 576px) {
-    max-width: 540px;
-  }
-
-  @media (min-width: 768px) {
-    max-width: 720px;
-  }
-
-  @media (min-width: 992px) {
-    max-width: 960px;
-  }
-
-  @media (min-width: 1200px) {
-    max-width: 1140px;
-  }
-`;
 
 const TitleContainer = styled.div`
   color: #ffffff;
